Article scaffolding: update node name generation algorithm
When article page is created using scaffolding, it is not showing all title for long title but cutting off. It needs to show all title, all in lowercase. Spaces should be replaced with ““, not “_”.
It is adding space in front of the page name, forcing all article to have “_” in the beginning. This should not happen.